adding TC-WBT-ADM002-01 - 04, fixing the 04 test scenario

+++ app/Http/Controllers/User/PegawaiController.php
@@ -76,7 +76,10 @@ class PegawaiController extends Controller
     // logic to remove a user
     public function destroy($user_id)
     {
-        $user = User::findOrFail($user_id); //
+        $user = User::findOrFail($user_id); 
+        if ($user->transaksis()->exists()) {
+            return back()->with('error', 'Pegawai tidak dapat dihapus karena memiliki transaksi aktif');
+        }
         $user->delete(); //
         return back()->with('success', 'Pegawai telah dihapus.');
     }
